Abstract S3 Compatible Storage Module logic
